Python/car_message/raspberry/requirements.txt
PySerial
requests
